[Refactor] UserDto, CategoryDto를 TaskResponseDto의 inner클래스로 수정
